add configuration doc for lambci lambda + some simplifications
