+++ pkg/lambci-lambda/util.go
@@ -2,18 +2,10 @@ package lambcilambda
 
 import (
 	"bytes"
-	"fmt"
-	"net/http"
 	"strconv"
 	"strings"
-	"time"
 
 	log "github.com/sirupsen/logrus"
-)
-
-const (
-	maxRetries = 3
-	retryDelay = 1 * time.Second
 )
 
 // determine the next version from a list of layer directories
@@ -56,46 +48,4 @@ func Map2Lines(m map[string]string) string {
 		lines.WriteString("\n")
 	}
 	return lines.String()
-}
-
-// post an HTTP request and return result
-func HttpPost(url, data string) (*bytes.Buffer, error) {
-
-	doPost := func() (*bytes.Buffer, error) {
-
-		response, err := http.Post(url, "application/json", strings.NewReader(data))
-		if err != nil {
-			return nil, err
-		}
-		defer response.Body.Close()
-
-		if response.StatusCode != http.StatusOK {
-			return nil, fmt.Errorf("POST %s (%s) returned status %s", url, data, response.Status)
-		}
-
-		result := new(bytes.Buffer)
-		_, err = result.ReadFrom(response.Body)
-		if err != nil {
-			return nil, err
-		}
-
-		return result, nil
-	}
-
-	var err error
-	var result *bytes.Buffer
-
-	retries := 0
-	for {
-
-		result, err = doPost()
-		if err == nil || retries >= maxRetries {
-			break
-		}
-
-		time.Sleep(retryDelay)
-		retries++
-	}
-
-	return result, err
 }
