+++ pkg/lambci-lambda/util_test.go
@@ -1,9 +1,6 @@
 package lambcilambda_test
 
 import (
-	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/serverlessresearch/srk/pkg/lambci-lambda"
@@ -21,23 +18,4 @@ func TestMap2Lines(t *testing.T) {
 
 	assert.Equal(t, "", lambcilambda.Map2Lines(nil))
 	assert.Equal(t, "key1=value1\nkey2=value2\n", lambcilambda.Map2Lines(map[string]string{"key1": "value1", "key2": "value2"}))
-}
-
-func TestHttpPost(t *testing.T) {
-
-	var received string
-	data := "hello, world"
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		data, _ := ioutil.ReadAll(r.Body)
-		received = string(data)
-		w.Write(data)
-	}))
-	defer ts.Close()
-
-	result, err := lambcilambda.HttpPost(ts.URL, data)
-	assert.Nil(t, err)
-
-	assert.Equal(t, data, received)
-	assert.Equal(t, data, result.String())
 }
